reorganize store files
