+++ gridsome/lib/store/__tests__/__fixtures__/JSONTransformer.js
